# docs/swagger.yaml
basePath: /atenea
definitions:
...
  models.RolParam:
    properties:
      id:
        type: integer
      nombre:
        type: string
    type: object
  models.Usuario:
    properties:
      cel:
        type: string
      email:
        type: string
      id:
        type: integer
      id_rol:
        type: integer
      nombre:
        type: string
      pass:
        type: string
      usuario:
        type: string
    type: object
  models.UsuarioAuthParam:
    properties:
      password:
        type: string
      usuario:
        type: string
    type: object
  models.UsuarioParam:
    properties:
      cel:
        type: string
      email:
        type: string
      id:
        type: integer
      id_rol:
        type: integer
      nombre:
        type: string
      password:
        type: string
      usuario:
        type: string
    type: object
host: IPServer:3021
info:
  contact:
    email: rgarciaxxx@yahoo.com.mx
    name: 'Soporte para las APIS: Raúl García'
    url: http://www.ineel.mx
  description: Contiene las APIs para atender las peticiones de la IHM hacia la base
    de datos ATENEA
  license:
    name: Atenea Inc.
    url: http://www.atenea.mx/LICENSE-2.0.html
  title: Servidor de APIs para ATENEA
  version: "1.0"
paths:
...
  /rol:
    post:
      consumes:
      - application/json
      description: Agrega un nuevo rol
      parameters:
      - description: Estructura del rol a agregar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.RolParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Crea un nuevo registro de rol
      tags:
      - rol
  /rol/id/{id}:
    delete:
      consumes:
      - application/json
      description: Borra el rol correspondiente al valor del identificador (id) indicado.
      parameters:
      - description: Identificador (id) del rol a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Elimina roles
      tags:
      - rol
  /usuario:
    post:
      consumes:
      - application/json
      description: Agrega un nuevo usuario
      parameters:
      - description: Estructura del usuario a agregar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.UsuarioParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Crea un nuevo registro de usuario
      tags:
      - usuario
    put:
      consumes:
      - application/json
      description: Actualiza el registro del usuario indicado en el body de la petición
      parameters:
      - description: Estructura del usuario a actualizar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.UsuarioParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Actualiza datos de usuarios
      tags:
      - usuario
  /usuario/actpass:
    put:
      consumes:
      - application/json
      description: Actualiza el password del usuario indicado en el body de la petición
      parameters:
      - description: Estructura del usuario y pwd a actualizar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.UsuarioAuthParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Actualiza el password de usuarios
      tags:
      - usuario
  /usuario/id/{id}:
    delete:
      consumes:
      - application/json
      description: Borra el usuario correspondiente al valor del identificador (id)
        indicado.
      parameters:
      - description: Identificador (id) del usuario a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Elimina usuarios
      tags:
      - usuario
  /usuario/valida:
    post:
      consumes:
      - application/json
      description: Valida credenciales del usuario indicado en el body de la petición
      parameters:
      - description: Estructura del usuario a agregar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.UsuarioAuthParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Usuario'
        "400":
          description: ""
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: ""
      summary: Valida credenciales de usuarios
      tags:
      - usuario
swagger: "2.0"